# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: MockApp API
  version: 1.2.0
  description: |
    Production-ready REST API for **MockApp**, a modern Flutter mobile application.

    - **Response envelope:** every JSON response uses `{ success, message, data | errors, … }`.
    - **Auth:** JWT access token (15 min target, 7 d in this mock) + 30-day refresh token.
    - **Pagination:** `page` / `limit` query params; `meta` block in response.
    - **Edge cases:** append `?scenario=error|validation|unauthorized|forbidden|not_found|maintenance|rate_limit|force_update|empty|slow|very_slow|timeout` to any endpoint.

    See `00-overview.md` for full conventions and rationale.

    **Fixed test accounts** (stable across reseeds):
    ```
    admin@mock.com / password123
    user@mock.com  / password123
    ```
  contact:
    name: MockApp Team
    email: dev@mockapp.com
    url: https://github.com/mockapp/mock-api-nextjs
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Local Next.js mock server (current code paths)
  - url: http://localhost:3001/api/v1
    description: Local Mockoon mock server (target /v1 paths)
  - url: https://mock.mockapp.com/api/v1
    description: Hosted mock (staging)
...
security:
  - BearerAuth: []

# ─────────────────────────────────────────────────────────────────────────────
# PATHS
# ─────────────────────────────────────────────────────────────────────────────
paths:
...
  /cart/items/{product_id}:
    parameters:
      - { name: product_id, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Cart]
      summary: Set the cart line quantity to an exact value
      description: A `quantity` of `0` removes the line (returns 204).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: { type: integer, minimum: 0 }
      responses:
        '200':
          description: Cart line updated
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponseBase'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/CartItem' }
        '204':
          description: Quantity set to 0 — line removed
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
    delete:
      tags: [Cart]
      summary: Remove a single line from the cart
      responses:
        '204':
          description: Line removed
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ─── ORDERS ───────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token in `Authorization: Bearer <token>`.
        Obtain via `POST /auth/login` or `POST /auth/register`.
        On `401 AUTH_TOKEN_EXPIRED`, refresh via `POST /auth/refresh`.

  # ─── Parameters ──────────────────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing, invalid, or expired access token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            expired:
              summary: Token expired
              value:
                success: false
                message: Your session has expired. Please sign in again.
                errors: null
                error_code: AUTH_TOKEN_EXPIRED
                request_id: req_01HF3XK2N7A8YBPQ4ZM
            missing:
              summary: Missing token
              value:
                success: false
                message: Authentication required.
                errors: null
                error_code: AUTH_TOKEN_INVALID
                request_id: req_01HF3XK2N7A8YBPQ4ZM
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            not_found:
              value:
                success: false
                message: The requested resource was not found.
                errors: null
                error_code: RESOURCE_NOT_FOUND
                request_id: req_01HF3XK2N7A8YBPQ4ZM
    ValidationError:
      description: Validation error (field-level details in `errors`)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          examples:
            validation:
              value:
                success: false
                message: The given data was invalid.
                errors:
                  email: ["The email field is required.", "The email must be valid."]
                  password: ["The password must be at least 8 characters."]
                error_code: VALIDATION_ERROR
                request_id: req_01HF3XK2N7A8YBPQ4ZM
...
  schemas:
...
    SuccessResponseBase:
      type: object
      required: [success, message, data]
      properties:
        success: { type: boolean, const: true }
        message: { type: string, example: Success }
        data: {}  # narrowed per endpoint
      additionalProperties: false
...
    ErrorResponse:
      type: object
      required: [success, message, errors, error_code, request_id]
      properties:
        success: { type: boolean, const: false }
        message:
          type: string
          example: The given data was invalid.
        errors:
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties:
                type: array
                items: { type: string }
          example:
            email: ["The email field is required."]
        error_code:
          type: string
          enum:
            - BAD_REQUEST
            - AUTH_INVALID_CREDENTIALS
            - AUTH_TOKEN_EXPIRED
            - AUTH_TOKEN_INVALID
            - AUTH_REFRESH_TOKEN_INVALID
            - AUTH_ACCOUNT_DISABLED
            - AUTHORIZATION_FORBIDDEN
            - RESOURCE_NOT_FOUND
            - RESOURCE_ALREADY_EXISTS
            - FILE_TOO_LARGE
            - UNSUPPORTED_MEDIA_TYPE
            - VALIDATION_ERROR
            - FORCE_UPDATE_REQUIRED
            - RATE_LIMIT_EXCEEDED
            - MAINTENANCE_MODE
            - INTERNAL_SERVER_ERROR
        request_id:
          type: string
          example: req_01HF3XK2N7A8YBPQ4ZM
...
    CartItem:
      type: object
      required: [id, product_id, product, quantity, unit_price, subtotal, currency]
      properties:
        id:         { type: string, format: uuid }
        product_id: { type: string, format: uuid }
        product:
          type: object
          properties:
            id:             { type: string, format: uuid }
            name:           { type: string }
            image:          { type: string, format: uri, nullable: true }
            price:          { type: number }
            discount_price: { type: number, nullable: true }
            stock:          { type: integer }
            is_active:      { type: boolean }
        quantity:   { type: integer, minimum: 1 }
        unit_price: { type: number, description: "Effective price (discount_price if set, otherwise price)" }
        subtotal:   { type: number, description: "unit_price × quantity, rounded" }
        currency:   { type: string, example: IDR }
        added_at:   { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...